Split _pql and _query methods (thx b4ldr)
and refactor the testsby splitting them into
more files too, like the API code.
