+++ pypuppetdb/api/__init__.py
@@ -1,3 +1,4 @@
+from pypuppetdb.api.pql import PqlAPI
 from pypuppetdb.api.query import QueryAPI
 from pypuppetdb.api.metrics import MetricsAPI
 from pypuppetdb.api.other import CmdAPI, MetaAPI, ServicesAPI
@@ -9,5 +10,6 @@ class API(
     CmdAPI,
     MetaAPI,
     ServicesAPI,
+    PqlAPI,
 ):
     pass
